# platform/docs/swagger.yaml
components:
  schemas:
...
    httpx.ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    underwriter.AllocationView:
      properties:
        nama:
          example: Mirae Asset Sekuritas
          type: string
        participant:
          description: Broker code the shares were credited to.
          example: YP
          type: string
        percentage:
          description: Share of listed_shares this tranche represents, to two decimals.
          example: 62.5
          type: number
        shares:
          description: Shares allocated.
          example: 25000
          format: int64
          type: integer
        underwriter:
          description: |-
            Broker code that took this tranche. Identical to participant, since an
            underwriter is a participant; both are present so the field a client keys
            on does not depend on knowing that.
          example: YP
          type: string
        value:
          description: 'Proceeds for this tranche: shares × ipo_price.'
          example: 25000000
          format: int64
          type: integer
      type: object
    underwriter.CreateRequest:
      properties:
        participant:
          description: |-
            Broker code permitted to underwrite. Must be a registered participant.
            Allocated shares land in this participant's holdings, because only a
            participant can trade them.
          example: YP
          type: string
      required:
      - participant
      type: object
    underwriter.ExistingIPORequest:
      properties:
        ipo_price:
          description: |-
            Offering price per share. Must be > 0; becomes the instrument's reference
            price until it first trades.
          example: 1000
          format: int64
          type: integer
        underwriters:
          description: |-
            The syndicate. At least one entry, and the tranches must sum to exactly
            the instrument's existing listed_shares. Every member is on equal terms.
          items:
            $ref: '#/components/schemas/underwriter.UnderwriterAllocation'
          type: array
          uniqueItems: false
      required:
      - ipo_price
      - underwriters
      type: object
    underwriter.IPORequest:
      properties:
        ipo_price:
          description: |-
            Offering price per share. Must be > 0; becomes the instrument's reference
            price until it first trades.
          example: 1000
          format: int64
          type: integer
        kode:
          description: Instrument code, unique across the exchange.
          example: SSIA
          type: string
        listed_shares:
          description: Publicly tradeable shares. The allocations must sum to exactly
            this.
          example: 40000
          format: int64
          type: integer
        nama:
          description: Company name.
          example: PT Surya Cipta Semesta Tbk
          type: string
        underwriters:
          description: |-
            The syndicate. At least one entry, and the tranches must sum to exactly
            listed_shares. Every member is on equal terms.
          items:
            $ref: '#/components/schemas/underwriter.UnderwriterAllocation'
          type: array
          uniqueItems: false
        unlisted_shares:
          description: Restricted shares not available for trading. Defaults to 0.
          example: 10000
          format: int64
          type: integer
      required:
      - ipo_price
      - kode
      - listed_shares
      - nama
      - underwriters
      type: object
    underwriter.IPOResponse:
      properties:
        ipo_price:
          description: Offering price per share.
          example: 1000
          format: int64
          type: integer
        kode:
          example: SSIA
          type: string
        listed_shares:
          example: 40000
          format: int64
          type: integer
        nama:
          example: PT Surya Cipta Semesta Tbk
          type: string
        total_shares:
          example: 50000
          format: int64
          type: integer
        total_value:
          description: 'Total raised: listed_shares × ipo_price.'
          example: 40000000
          format: int64
          type: integer
        underwriters:
          description: The syndicate, lead first.
          items:
            $ref: '#/components/schemas/underwriter.AllocationView'
          type: array
          uniqueItems: false
        unlisted_shares:
          example: 10000
          format: int64
          type: integer
      type: object
    underwriter.UnderwriterAllocation:
      properties:
        shares:
          description: Shares allocated. Must be > 0.
          example: 25000
          format: int64
          type: integer
        underwriter:
          description: Broker code taking this tranche. Must be a registered underwriter.
          example: YP
          type: string
      required:
      - shares
      - underwriter
      type: object
    underwriter.UnderwriterView:
      properties:
        is_active:
          example: true
          type: boolean
        kode:
          description: |-
            Broker code. Identical to participant; both are present so a client can
            key on kode like every other listing without knowing they are the same.
          example: YP
          type: string
        nama:
          example: Mirae Asset Sekuritas
          type: string
        participant:
          description: Broker code this underwriter trades through.
          example: YP
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Static administrative key, from configuration. Guards every /api/admin route.
      in: header
      name: X-API-Key
      type: apiKey
    ParticipantKeyAuth:
      description: Per-broker key, issued by an admin and stored hashed. Guards every /api/participant route. Returned once at issue and never retrievable again.
      in: header
      name: X-Participant-Key
      type: apiKey
...
info:
  description: |-
    REST surface for the JAST core matching engine (JATS-style continuous
    matching for Bursa Efek Indonesia). The engine knows only brokers
    (participants), not individual investors, and stops at trade execution.


    There are two authentication tiers, and they never mix:

    - **Participant** (`/api/participant/*`, `/ws/participant/*`) — a broker
    authenticates with its own key, stored hashed in the database and sent as
    `X-Participant-Key`. Keys are issued and revoked per broker at runtime.
    - **Admin** (`/api/admin/*`, `/ws/admin/*`) — the single static key from
    configuration, sent as `X-API-Key`.

    Orders enter only through POST /api/participant/orders. The order book can
    be polled or streamed over WebSocket (documented, but not exercisable from
    Swagger UI, since browsers cannot set headers on a WS handshake).
  title: JAST Core API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/admin/emiten/{kode}/ipo:
    post:
      description: |-
        Takes an instrument that was registered through POST /api/admin/emiten —
        and is therefore still **dormant** — public: it hands the shares to the
        syndicate and opens the order book. This is the only way an instrument
        becomes tradeable.

        The body carries no share counts. listed_shares and unlisted_shares are the
        instrument's own, fixed when it was registered — an offering decides who
        underwrites it and at what price, not how many shares the company has. The
        tranches must sum to exactly the listed_shares it already carries.

        The syndicate rules are the same as POST /api/admin/ipo:
        - each `underwriter` is a **broker code** that has been permitted to underwrite
        - the tranches must sum to **exactly** the instrument's listed_shares
        - every member is on equal terms; there is no lead

        ipo_price becomes the instrument's reference price until it first trades. An
        instrument that is already trading is rejected — a second offering would
        issue its shares twice.
      operationId: runIPOForEmiten
      parameters:
      - description: Emiten code
        example: BBNI
        in: path
        name: kode
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/underwriter.ExistingIPORequest'
              description: Offering to run
              summary: body
        description: Offering to run
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/underwriter.IPOResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Bad syndicate, shares do not sum, unknown underwriter, or emiten
            already trading
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Unknown emiten
      security:
      - ApiKeyAuth: []
      summary: Run an IPO over a registered instrument
      tags:
      - admin
...
  /api/admin/ipo:
    post:
      description: |-
        Lists an instrument and hands its shares to the underwriting syndicate in
        one step — an instrument whose shares sit nowhere is not yet an offering.

        The syndicate is flat — every member takes a tranche on equal terms:
        - each `underwriter` is a **broker code** that has been permitted to underwrite
        - the tranches must sum to **exactly** listed_shares — a short sum leaves
        shares in nobody's hands, an over-sum conjures shares that were never issued

        ipo_price becomes the instrument's reference price until it first trades, so
        it has a valuation and something to quote against from the moment it lists.
        The allocation is one transaction: the audit rows and the share credits move
        together or not at all.
      operationId: runIPO
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/underwriter.IPORequest'
              description: Offering to run
              summary: body
        description: Offering to run
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/underwriter.IPOResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Bad syndicate, shares do not sum, or unknown underwriter
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Instrument code already exists
      security:
      - ApiKeyAuth: []
      summary: Run an IPO
      tags:
      - admin
...
  /api/admin/underwriters:
    get:
      description: |-
        Every broker permitted to underwrite offerings, ordered by code.

        An underwriter is a participant with a permission, not a firm of its own,
        so kode and nama are that participant's — `kode` and `participant` are
        always the same value. Allocated shares land in that broker's holdings,
        because only a participant can trade them.

        Members carry no rank: every underwriter in a syndicate is on equal terms.
      operationId: listUnderwriters
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/underwriter.UnderwriterView'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
      security:
      - ApiKeyAuth: []
      summary: List underwriters
      tags:
      - admin
    post:
      description: |-
        Grants an existing broker the right to underwrite offerings. That is all an
        underwriter is — a participant with a permission — so the request carries
        only the broker code: its name and code are the participant's already, and
        are joined on read rather than stored twice.

        The broker is also its trading identity: IPO allocations credit that
        participant's holdings, so an underwriter without one could be handed
        shares it could never sell.

        A broker may be registered once. Registering it again is a 409.
      operationId: createUnderwriter
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/underwriter.CreateRequest'
              description: Broker to permit
              summary: body
        description: Broker to permit
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/underwriter.UnderwriterView'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing participant, or unknown participant
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Missing or wrong X-API-Key
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.ErrorResponse'
          description: Broker is already an underwriter
      security:
      - ApiKeyAuth: []
      summary: Permit a broker to underwrite
      tags:
      - admin
...
servers:
- description: Local development
  url: http://localhost:8080